requirements.txt
cudatoolkit=9.2
imageio=2.9.0=py_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
matplotlib=3.3.3=py37h89c1867_0
numpy=1.19.1=py37hbc911f0_0
pandas=1.2.0=py37ha9443f7_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pydicom=2.1.2=pyhd3deb0d_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda9.2.148_cudnn7.6.3_0
regex=2021.4.4=py37h5e8e339_0
requests=2.25.1=pyhd3deb0d_0
scikit-learn=0.24.1=py37h69acf81_0
scipy=1.6.0=py37h14a347d_0
setuptools=51.3.3=py37h06a4308_4
simpleitk=2.0.2=py37h3fd9d12_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.1=py37hcb7a40c_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu92